requirements.txt
Flask
Flask-CORS
xgboost
pandas
scikit-learn
shap
mysql-connector-python
gunicorn
textblob